BrahmNET/Pinecone_PersistantMemory_RAG/requirements.txt
langchain==0.3.26
langchain-community==0.3.27
langchain-cohere==0.4.4
langchain-groq
cohere==5.14.0
streamlit==1.26.0
pinecone-client==3.0.0
beautifulsoup4
pypdf
